requirements.txt
python-dotenv==1.0.1
streamlit==1.39.0
streamlit-card==1.0.2
streamlit-extras==0.5.0
litellm==1.51.1
langchain-community==0.3.3
langchain-openai==0.2.4
langchain-anthropic==0.3.7
langchain==0.2.0
tavily-python==0.3.1